+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Changelog for NeoFS Node
 - neofs-adm uses alphabet0/1/2 names for wallet files by default now (#3268)
 - Distribute load in writecache during flushing (#3261)
 - Faster migration in peapod-to-fstree (#3280)
+- `peapod-to-fstree` migrates only objects with existing containers (#3283)
 
 ### Removed
 - SN `apiclient.allow_external` config (#3235)